set up new test structure for textarea
