+++ packages/textarea/package.json
@@ -21,9 +21,9 @@
   ],
   "scripts": {
     "build": "microbundle --name Textarea",
-    "test": "jest --coverage",
-    "dev": "webpack-dev-server --config example/webpack.config.js",
-    "prod": "webpack --config example/webpack.config.js --mode production",
+    "test": "jest --coverage & npx playwright test",
+    "dev": "webpack-dev-server --config tools/webpack.config.js",
+    "prod": "webpack --config tools/webpack.config.js --mode production",
     "prepublish": "npm run -s build"
   },
   "gitHead": "9042db4005411c4360ba3e8188e03333c10098f2"
